Run by line partial implementation (microsoft/vscode-python#11608)
* Preliminary getting variables from the debugger

* Implement updating as stepping

* Get debugger and regular kernel to use same scripts

* Put back old code and create experiment to enable/disable this

* Implement experiment and begin on test for debugger

* Finish fixing tests

* Add news entry

* Create multiplexer

* Basic idea of glyph updating and control

* Forget the svgs

* Rework to go through interactive path

* Highlight line kinda showing up

* Fix variables to come out.

* Fix ip line sticking around too long

* Colors and icon working for ip

* Implement F10 and F5

* Hover and step into icons

* Remove unnecessary variables and support step into

* Update frame id for variable

* Fixup after merge

* Exclude everything glob for just my code

* Try fixing order for messages to debugger
Remove exclusions as they mess up subsequent runs

* Fix request problem and dark theme

* Add news entry

* Some extra comments and fix sonar problems

* Add loc strings and make constants for code icons

* Sonar problems and removing unused code

* Change name to debugInfo

* Try again to exclude the codicon files

* Wrong base directory

* Fixes from code review and run by line new icon

* Fix functional test failures

* Fix unit test failure (yay, caught a bug)

* Fix stupid sonar warnings

* Fix some more failing tests

+++ extensions/positron-python/src/test/datascience/raw-kernel/rawKernel.functional.test.ts
@@ -95,11 +95,13 @@ suite('DataScience raw kernel tests', () => {
     }
 
     async function disconnectFromKernel() {
-        await kernelProcess.dispose().catch(noop);
-        try {
-            await fs.remove(connectionFile);
-        } catch {
-            noop();
+        if (kernelProcess) {
+            await kernelProcess.dispose().catch(noop);
+            try {
+                await fs.remove(connectionFile);
+            } catch {
+                noop();
+            }
         }
     }
 
@@ -111,7 +113,7 @@ suite('DataScience raw kernel tests', () => {
         let exited = false;
         kernelProcess.exited(() => (exited = true));
         await shutdown();
-        await sleep(500); // Give time for the shutdown to go across
+        await sleep(2500); // Give time for the shutdown to go across
         assert.ok(exited, 'Kernel did not shutdown');
     });
